482 neon cli refactoring (#483)
* #482_neon_cli_refactoring

* #482_neon_cli_refactoring

* #482_neon_cli_refactoring
