+++ evm_loader/cli/src/commands/mod.rs
@@ -1,0 +1,9 @@
+pub mod cancel_trx;
+pub mod create_ether_account;
+pub mod create_program_address;
+pub mod deploy;
+pub mod emulate;
+pub mod get_ether_account_data;
+pub mod get_neon_elf;
+pub mod get_storage_at;
+pub mod update_valids_table;
