+++ evm_loader/cli/src/main.rs
@@ -5,47 +5,44 @@ mod account_storage;
 
 mod errors;
 mod logs;
+mod commands;
 
 use crate::{
     account_storage::{
         make_solana_program_address,
-        EmulatorAccountStorage,
-        AccountJSON,
-        TokenAccountJSON,
+    },
+    commands::{
+        emulate,
+        create_program_address,
+        create_ether_account,
+        deploy,
+        get_ether_account_data,
+        cancel_trx,
+        get_neon_elf,
+        get_storage_at,
+        update_valids_table,
     },
 };
 
 use evm_loader::{
-    instruction::EvmInstruction,
     account_data::{
         ACCOUNT_SEED_VERSION,
         AccountData,
-        Account,
-        Contract
     },
     config::{  collateral_pool_base },
 };
 
-use evm::{H160, H256, U256, ExitReason,};
+use evm::{H160, H256, U256};
 use solana_sdk::{
-    clock::Slot,
     commitment_config::{CommitmentConfig, CommitmentLevel},
-    incinerator,
-    instruction::{AccountMeta, Instruction},
+    instruction::{Instruction},
     message::Message,
     pubkey::Pubkey,
     signature::{Keypair, Signer, Signature},
-    signers::Signers,
     keccak::Hasher,
     transaction::Transaction,
-    system_program,
-    sysvar,
     system_instruction,
-    account_utils::StateMut,
-    bpf_loader, bpf_loader_deprecated,
-    bpf_loader_upgradeable::{self, UpgradeableLoaderState},
 };
-use serde_json::json;
 use std::{
     collections::HashMap,
     io::{Read},
@@ -54,13 +51,9 @@ use std::{
     str::FromStr,
     process::exit,
     sync::Arc,
-    thread::sleep,
-    time::{Duration},
-    convert::{TryFrom, TryInto},
+    convert::{TryInto},
     fmt,
     fmt::{Debug, Display,},
-    cell::RefCell,
-    rc::Rc
 };
 
 use clap::{
@@ -80,44 +73,15 @@ use solana_clap_utils::{
 
 use solana_client::{
     rpc_client::RpcClient,
-    rpc_config::{RpcSendTransactionConfig, RpcTransactionConfig},
-    rpc_request::MAX_GET_SIGNATURE_STATUSES_QUERY_ITEMS,
-    tpu_client::{TpuClient, TpuClientConfig},
+    rpc_config::{RpcSendTransactionConfig},
 };
-use solana_cli::{
-    checks::{check_account_for_fee},
-};
-use solana_cli_output::display::new_spinner_progress_bar;
-use solana_transaction_status::{
-    TransactionConfirmationStatus,
-    UiTransactionEncoding,
-    EncodedTransaction,
-    UiMessage,
-    UiInstruction,
-    EncodedConfirmedTransaction
-};
-
-use libsecp256k1::SecretKey;
-use libsecp256k1::PublicKey;
 
 use rlp::RlpStream;
 
-use log::{debug, error, info};
+use log::{debug, error};
 use logs::LogContext;
 
-use crate::account_storage::SolanaAccountJSON;
-use evm_loader::{
-    executor_state::{
-        ExecutorState,
-        ExecutorSubstate,
-    },
-    executor::Machine,
-    solana_backend::AccountStorage,
-    solidity_account::SolidityAccount
-};
 use crate::errors::NeonCliError;
-
-const DATA_CHUNK_SIZE: usize = 229; // Keep program chunks under PACKET_DATA_SIZE
 
 type Error = Box<dyn std::error::Error>;
 type CommandResult = Result<(), Error>;
@@ -139,204 +103,6 @@ impl Debug for Config {
     }
 }
 
-#[allow(clippy::too_many_lines)]
-fn command_emulate(config: &Config, contract_id: Option<H160>, caller_id: H160, data: Option<Vec<u8>>,
-                   value: Option<U256>, token_mint: &Pubkey) -> CommandResult {
-    debug!("command_emulate(config={:?}, contract_id={:?}, caller_id={:?}, data={:?}, value={:?})",
-        config,
-        contract_id,
-        caller_id,
-        &hex::encode(data.clone().unwrap_or_default()),
-        value);
-
-    let storage = match &contract_id {
-        Some(program_id) =>  {
-            debug!("program_id to call: {:?}", *program_id);
-            EmulatorAccountStorage::new(config, *program_id, caller_id, *token_mint)
-        },
-        None => {
-            let (solana_address, _nonce) = make_solana_program_address(&caller_id, &config.evm_loader);
-            let trx_count = get_ether_account_nonce(config, &solana_address, token_mint)?;
-            let trx_count= trx_count.0;
-            let program_id = get_program_ether(&caller_id, trx_count);
-            debug!("program_id to deploy: {:?}", program_id);
-            EmulatorAccountStorage::new(config, program_id, caller_id, *token_mint)
-        }
-    };
-
-    let (exit_reason, result, applies_logs, used_gas, steps_executed) = {
-        // u64::MAX is too large, remix gives this error:
-        // Gas estimation errored with the following message (see below).
-        // Number can only safely store up to 53 bits
-        let gas_limit = 50_000_000;
-        let executor_substate = Box::new(ExecutorSubstate::new(gas_limit, &storage));
-        let executor_state = ExecutorState::new(executor_substate, &storage);
-        let mut executor = Machine::new(executor_state);
-        debug!("Executor initialized");
-
-        let (result, exit_reason) = match &contract_id {
-            Some(_) =>  {
-                debug!("call_begin(storage.origin()={:?}, storage.contract()={:?}, data={:?}, value={:?})",
-                    storage.origin(),
-                    storage.contract(),
-                    &hex::encode(data.clone().unwrap_or_default()),
-                    value);
-                executor.call_begin(storage.origin(),
-                                    storage.contract(),
-                                    data.unwrap_or_default(),
-                                    value.unwrap_or_default(),
-                                    gas_limit)?;
-                executor.execute()
-            },
-            None => {
-                debug!("create_begin(storage.origin()={:?}, data={:?}, value={:?})",
-                    storage.origin(),
-                    &hex::encode(data.clone().unwrap_or_default()),
-                    value);
-                executor.create_begin(storage.origin(),
-                                      data.unwrap_or_default(),
-                                      value.unwrap_or_default(),
-                                      gas_limit)?;
-                executor.execute()
-            }
-        };
-        debug!("Execute done, exit_reason={:?}, result={:?}", exit_reason, result);
-        debug!("{} steps executed", executor.get_steps_executed());
-
-        let steps_executed = executor.get_steps_executed();
-        let executor_state = executor.into_state();
-        let used_gas = executor_state.gasometer().used_gas() + 1; // "+ 1" because of https://github.com/neonlabsorg/neon-evm/issues/144
-        let refunded_gas = executor_state.gasometer().refunded_gas();
-        let needed_gas = used_gas + (if refunded_gas > 0 { u64::try_from(refunded_gas)? } else { 0 });
-        debug!("used_gas={:?} refunded_gas={:?}", used_gas, refunded_gas);
-        if exit_reason.is_succeed() {
-            debug!("Succeed execution");
-            let apply = executor_state.deconstruct();
-            (exit_reason, result, Some(apply), needed_gas, steps_executed)
-        } else {
-            (exit_reason, result, None, needed_gas, steps_executed)
-        }
-    };
-
-    debug!("Call done");
-    let status = match exit_reason {
-        ExitReason::Succeed(_) => {
-            let (applies, _logs, transfers, spl_transfers, spl_approves, erc20_approves) = applies_logs.unwrap();
-
-            storage.apply(applies);
-            storage.apply_transfers(transfers, token_mint);
-            storage.apply_spl_approves(spl_approves);
-            storage.apply_spl_transfers(spl_transfers);
-            storage.apply_erc20_approves(erc20_approves);
-
-            debug!("Applies done");
-            "succeed".to_string()
-        }
-        ExitReason::Error(_) => "error".to_string(),
-        ExitReason::Revert(_) => "revert".to_string(),
-        ExitReason::Fatal(_) => "fatal".to_string(),
-        ExitReason::StepLimitReached => unreachable!(),
-    };
-
-    info!("{}", &status);
-    info!("{}", &hex::encode(&result));
-
-    if !exit_reason.is_succeed() {
-        debug!("Not succeed execution");
-    }
-
-    let accounts: Vec<AccountJSON> = storage.get_used_accounts();
-
-    let solana_accounts: Vec<SolanaAccountJSON> = storage.solana_accounts
-        .borrow()
-        .values()
-        .cloned()
-        .map(SolanaAccountJSON::from)
-        .collect();
-
-    let token_accounts: Vec<TokenAccountJSON> = storage.token_accounts
-        .borrow()
-        .values()
-        .cloned()
-        .map(TokenAccountJSON::from)
-        .collect();
-
-    let js = json!({
-        "accounts": accounts,
-        "solana_accounts": solana_accounts,
-        "token_accounts": token_accounts,
-        "result": &hex::encode(&result),
-        "exit_status": status,
-        "exit_reason": exit_reason,
-        "used_gas": used_gas,
-        "steps_executed": steps_executed,
-    }).to_string();
-
-    println!("{}", js);
-
-    Ok(())
-}
-
-fn command_create_program_address (
-    config: &Config,
-    ether_address: &H160,
-) {
-    let (solana_address, nonce) = make_solana_program_address(ether_address, &config.evm_loader);
-    println!("{} {}", solana_address, nonce);
-}
-
-fn command_create_ether_account (
-    config: &Config,
-    ether_address: &H160,
-    lamports: u64,
-    space: u64,
-    token_mint: &Pubkey
-) -> CommandResult {
-
-    let (solana_address, nonce) = make_solana_program_address(ether_address, &config.evm_loader);
-    let token_address = spl_associated_token_account::get_associated_token_address(&solana_address, token_mint);
-    debug!("Create ethereum account {} <- {} {}", solana_address, hex::encode(ether_address), nonce);
-
-    let instruction = Instruction::new_with_bincode(
-            config.evm_loader,
-            &EvmInstruction::CreateAccount {lamports, space, ether: *ether_address, nonce},
-            vec![
-                AccountMeta::new(config.signer.pubkey(), true),
-                AccountMeta::new(solana_address, false),
-                AccountMeta::new(token_address, false),
-                AccountMeta::new_readonly(system_program::id(), false),
-                AccountMeta::new_readonly(*token_mint, false),
-                AccountMeta::new_readonly(spl_token::id(), false),
-                AccountMeta::new_readonly(spl_associated_token_account::id(), false),
-                AccountMeta::new_readonly(sysvar::rent::id(), false),
-            ]);
-
-    let finalize_message = Message::new(&[instruction], Some(&config.signer.pubkey()));
-    let (blockhash, fee_calculator) = config.rpc_client.get_recent_blockhash()?;
-
-    check_account_for_fee(
-        &config.rpc_client,
-        &config.signer.pubkey(),
-        &fee_calculator,
-        &finalize_message)?;
-
-    let mut finalize_tx = Transaction::new_unsigned(finalize_message);
-
-    finalize_tx.try_sign(&[&*config.signer], blockhash)?;
-    debug!("signed: {:x?}", finalize_tx);
-
-    config.rpc_client.send_and_confirm_transaction_with_spinner(&finalize_tx)?;
-
-    info!("{}", json!({
-        "solana": solana_address.to_string(),
-        "token": token_address.to_string(),
-        "ether": hex::encode(ether_address),
-        "nonce": nonce,
-    }));
-
-    Ok(())
-}
-
 fn read_program_data(program_location: &str) -> Result<Vec<u8>, Box<dyn std::error::Error>> {
     let mut file = File::open(program_location).map_err(|err| {
         format!("Unable to open program file '{}': {}", program_location, err)
@@ -347,134 +113,6 @@ fn read_program_data(program_location: &str) -> Result<Vec<u8>, Box<dyn std::err
     })?;
 
     Ok(program_data)
-}
-
-#[allow(clippy::too_many_lines)]
-fn send_and_confirm_transactions_with_spinner<T: Signers>(
-    rpc_client: &Arc<RpcClient>,
-    websocket_url: &str,
-    mut transactions: Vec<Transaction>,
-    signer_keys: &T,
-    commitment: CommitmentConfig,
-    mut last_valid_slot: Slot,
-) -> CommandResult {
-    let progress_bar = new_spinner_progress_bar();
-    let mut send_retries = 5;
-
-    progress_bar.set_message("Finding leader nodes...");
-    let tpu_client = TpuClient::new(
-        rpc_client.clone(),
-        websocket_url,
-        TpuClientConfig::default(),
-    )?;
-
-    loop {
-        // Send all transactions
-        let mut pending_transactions = HashMap::new();
-        let num_transactions = transactions.len();
-        for transaction in transactions {
-            if !tpu_client.send_transaction(&transaction) {
-                let _result = rpc_client
-                    .send_transaction_with_config(
-                        &transaction,
-                        RpcSendTransactionConfig {
-                            preflight_commitment: Some(commitment.commitment),
-                            ..RpcSendTransactionConfig::default()
-                        },
-                    )
-                    .ok();
-            }
-            pending_transactions.insert(transaction.signatures[0], transaction);
-            progress_bar.set_message(&format!(
-                "[{}/{}] Transactions sent",
-                pending_transactions.len(),
-                num_transactions
-            ));
-
-            // Throttle transactions to about 100 TPS
-            sleep(Duration::from_millis(10));
-        }
-
-        // Collect statuses for all the transactions, drop those that are confirmed
-        loop {
-            let mut slot = 0;
-            let pending_signatures = pending_transactions.keys().copied().collect::<Vec<_>>();
-            for pending_signatures_chunk in
-                pending_signatures.chunks(MAX_GET_SIGNATURE_STATUSES_QUERY_ITEMS)
-            {
-                if let Ok(result) = rpc_client.get_signature_statuses(pending_signatures_chunk) {
-                    let statuses = result.value;
-                    for (signature, status) in
-                        pending_signatures_chunk.iter().zip(statuses.into_iter())
-                    {
-                        if let Some(status) = status {
-                            if let Some(confirmation_status) = &status.confirmation_status {
-                                if *confirmation_status != TransactionConfirmationStatus::Processed
-                                {
-                                    pending_transactions.remove(signature);
-                                }
-                            } else if status.confirmations.is_none()
-                                || status.confirmations.unwrap() > 1
-                            {
-                                pending_transactions.remove(signature);
-                            }
-                        }
-                    }
-                }
-
-                slot = rpc_client.get_slot()?;
-                progress_bar.set_message(&format!(
-                    "[{}/{}] Transactions confirmed. Retrying in {} slots",
-                    num_transactions - pending_transactions.len(),
-                    num_transactions,
-                    last_valid_slot.saturating_sub(slot)
-                ));
-            }
-
-            if pending_transactions.is_empty() {
-                return Ok(());
-            }
-
-            if slot > last_valid_slot {
-                break;
-            }
-
-            for transaction in pending_transactions.values() {
-                if !tpu_client.send_transaction(transaction) {
-                    let _result = rpc_client
-                        .send_transaction_with_config(
-                            transaction,
-                            RpcSendTransactionConfig {
-                                preflight_commitment: Some(commitment.commitment),
-                                ..RpcSendTransactionConfig::default()
-                            },
-                        )
-                        .ok();
-                }
-            }
-
-            if cfg!(not(test)) {
-                // Retry twice a second
-                sleep(Duration::from_millis(500));
-            }
-        }
-
-        if send_retries == 0 {
-            return Err("Transactions failed".into());
-        }
-        send_retries -= 1;
-
-        // Re-sign any failed transactions with a new blockhash and retry
-        let (blockhash, _fee_calculator, new_last_valid_slot) = rpc_client
-            .get_recent_blockhash_with_commitment(commitment)?
-            .value;
-        last_valid_slot = new_last_valid_slot;
-        transactions = vec![];
-        for (_, mut transaction) in pending_transactions {
-            transaction.try_sign(signer_keys, blockhash)?;
-            transactions.push(transaction);
-        }
-    }
 }
 
 #[must_use]
@@ -519,97 +157,6 @@ impl rlp::Encodable for UnsignedTransaction {
         s.append_empty_data();
         s.append_empty_data();
     }
-}
-
-fn make_deploy_ethereum_transaction(
-    trx_count: u64,
-    program_data: &[u8],
-    caller_private: &SecretKey,
-    chain_id: u64
-) -> Vec<u8> {
-    let rlp_data = {
-        let tx = UnsignedTransaction {
-            to: None,
-            nonce: trx_count,
-            gas_limit: 9_999_999.into(),
-            gas_price: 0.into(),
-            value: 0.into(),
-            data: program_data.to_owned(),
-            chain_id: chain_id.into(),
-        };
-
-        rlp::encode(&tx).to_vec()
-    };
-
-    let (sig, rec) = {
-        use libsecp256k1::{Message, sign};
-        let msg = Message::parse(&keccak256(rlp_data.as_slice()));
-        sign(&msg, caller_private)
-    };
-
-    let mut msg : Vec<u8> = Vec::new();
-    msg.extend(sig.serialize().iter().copied());
-    msg.push(rec.serialize());
-    msg.extend((rlp_data.len() as u64).to_le_bytes().iter().copied());
-    msg.extend(rlp_data);
-
-    msg
-}
-
-fn fill_holder_account(
-    config: &Config,
-    holder: &Pubkey,
-    holder_id: u64,
-    msg: &[u8],
-) -> Result<(), Error> {
-    let creator = &config.signer;
-    let signers = [&*config.signer];
-
-    // Write code to holder account
-    debug!("Write code");
-    let mut write_messages = vec![];
-    for (chunk, i) in msg.chunks(DATA_CHUNK_SIZE).zip(0..) {
-        let offset = u32::try_from(i*DATA_CHUNK_SIZE)?;
-
-        let instruction = Instruction::new_with_bincode(
-            config.evm_loader,
-            /* &EvmInstruction::WriteHolder {holder_id, offset, bytes: chunk}, */
-            &(0x12_u8, holder_id, offset, chunk),
-            vec![AccountMeta::new(*holder, false),
-                 AccountMeta::new(creator.pubkey(), true)]
-        );
-
-        let message = Message::new(&[instruction], Some(&creator.pubkey()));
-        write_messages.push(message);
-    }
-    debug!("Send write message");
-
-    // Send write message
-    {
-        let (blockhash, _, last_valid_slot) = config.rpc_client
-            .get_recent_blockhash_with_commitment(CommitmentConfig::confirmed())?
-            .value;
-
-        let mut write_transactions = vec![];
-        for message in write_messages {
-            let mut tx = Transaction::new_unsigned(message);
-            tx.try_sign(&signers, blockhash)?;
-            write_transactions.push(tx);
-        }
-
-        debug!("Writing program data");
-        send_and_confirm_transactions_with_spinner(
-            &config.rpc_client,
-            &config.websocket_url,
-            write_transactions,
-            &signers,
-            CommitmentConfig::confirmed(),
-            last_valid_slot,
-        ).map_err(|err| format!("Data writes to program account failed: {}", err))?;
-        debug!("Writing program data done");
-    }
-
-    Ok(())
 }
 
 // fn get_ethereum_caller_credentials(
@@ -675,92 +222,6 @@ fn get_program_ether(
     keccak256_h256(&stream.out()).into()
 }
 
-fn get_ethereum_contract_account_credentials(
-    config: &Config,
-    caller_ether: &H160,
-    trx_count: u64,
-    token_mint: &Pubkey
-) -> (Pubkey, H160, u8, Pubkey, Pubkey, String) {
-    let creator = &config.signer;
-
-    let (program_id, program_ether, program_nonce) = {
-        let ether = get_program_ether(caller_ether, trx_count);
-        let (address, nonce) = make_solana_program_address(&ether, &config.evm_loader);
-        (address, ether, nonce)
-    };
-    debug!("Create account: {} with {} {}", program_id, program_ether, program_nonce);
-
-    let program_token = spl_associated_token_account::get_associated_token_address(&program_id, token_mint);
-
-    let (program_code, program_seed) = {
-        let seed: &[u8] = &[ &[ACCOUNT_SEED_VERSION], program_ether.as_bytes() ].concat();
-        let seed = bs58::encode(seed).into_string();
-        debug!("Code account seed {} and len {}", &seed, &seed.len());
-        let address = Pubkey::create_with_seed(&creator.pubkey(), &seed, &config.evm_loader).unwrap();
-        (address, seed)
-    };
-    debug!("Create code account: {}", &program_code.to_string());
-
-    (program_id, program_ether, program_nonce, program_token, program_code, program_seed)
-}
-
-#[allow(clippy::too_many_arguments)]
-fn create_ethereum_contract_accounts_in_solana(
-    config: &Config,
-    program_id: &Pubkey,
-    program_ether: &H160,
-    program_nonce: u8,
-    program_token: &Pubkey,
-    program_code: &Pubkey,
-    program_seed: &str,
-    program_code_len: usize,
-    token_mint: &Pubkey
-) -> Result<Vec<Instruction>, Error> {
-    let account_header_size = 1+Account::SIZE;
-    let contract_header_size = 1+Contract::SIZE;
-
-    let creator = &config.signer;
-    let program_code_acc_len = contract_header_size + program_code_len + 2*1024;
-
-    let minimum_balance_for_account = config.rpc_client.get_minimum_balance_for_rent_exemption(account_header_size)?;
-    let minimum_balance_for_code = config.rpc_client.get_minimum_balance_for_rent_exemption(program_code_acc_len)?;
-
-    if let Some(_account) = config.rpc_client.get_account_with_commitment(program_id, CommitmentConfig::confirmed())?.value
-    {
-        return Err("Account already exist".to_string().into());
-        // debug!("Account already exist");
-    }
-
-    let instructions = vec![
-        system_instruction::create_account_with_seed(
-            &creator.pubkey(),
-            program_code,
-            &creator.pubkey(),
-            program_seed,
-            minimum_balance_for_code,
-            program_code_acc_len as u64,
-            &config.evm_loader
-        ),
-        Instruction::new_with_bincode(
-            config.evm_loader,
-            &(2_u32, minimum_balance_for_account, 0_u64, program_ether.as_fixed_bytes(), program_nonce),
-            vec![
-                AccountMeta::new(creator.pubkey(), true),
-                AccountMeta::new(*program_id, false),
-                AccountMeta::new(*program_token, false),
-                AccountMeta::new(*program_code, false),
-                AccountMeta::new_readonly(system_program::id(), false),
-                AccountMeta::new_readonly(*token_mint, false),
-                AccountMeta::new_readonly(spl_token::id(), false),
-                AccountMeta::new_readonly(spl_associated_token_account::id(), false),
-                AccountMeta::new_readonly(sysvar::rent::id(), false),
-            ]
-        )
-    ];
-
-    Ok(instructions)
-}
-
 fn create_storage_account(config: &Config) -> Result<Pubkey, Error> {
     use rand::Rng;
     let mut rng = rand::thread_rng();
@@ -780,34 +241,6 @@ fn get_collateral_pool_account_and_index(config: &Config, collateral_pool_base: 
         &config.evm_loader).unwrap();
 
     (collateral_pool_account, collateral_pool_index)
-}
-
-fn parse_transaction_reciept(config: &Config, result: EncodedConfirmedTransaction) -> Option<Vec<u8>> {
-    let mut return_value : Option<Vec<u8>> = None;
-    if let EncodedTransaction::Json(transaction) = result.transaction.transaction {
-        if let UiMessage::Raw(message) = transaction.message {
-            let evm_loader_index = message.account_keys.iter().position(|x| *x == config.evm_loader.to_string());
-            if let Some(meta) = result.transaction.meta {
-                if let Some(inner_instructions) = meta.inner_instructions {
-                    for instruction in inner_instructions {
-                        if instruction.index == 0 {
-                            if let Some(UiInstruction::Compiled(compiled_instruction)) = instruction.instructions.iter().last() {
-                                if compiled_instruction.program_id_index as usize == evm_loader_index.unwrap() {
-                                    let decoded = bs58::decode(compiled_instruction.data.clone()).into_vec().unwrap();
-                                    if decoded[0] == 6 {
-                                        debug!("success");
-                                        return_value = Some(decoded[1..].to_vec());
-                                        break;
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    return_value
 }
 
 fn create_account_with_seed(
@@ -879,465 +312,6 @@ fn generate_random_holder_seed() -> (u64, String) {
     hasher.hash(&id.to_be_bytes()[bytes_count-significant_bytes_count..]);
     let output = hasher.result();
     (id, hex::encode(output)[..32].into())
-}
-
-#[allow(clippy::too_many_lines)]
-fn command_deploy(
-    config: &Config,
-    program_location: &str,
-    token_mint: &Pubkey,
-    collateral_pool_base: &Pubkey,
-    chain_id: u64
-) -> CommandResult {
-    let creator = &config.signer;
-    let program_data = read_program_data(program_location)?;
-    let operator_token = spl_associated_token_account::get_associated_token_address(&creator.pubkey(), token_mint);
-
-    // Create ethereum caller private key from sign of array by signer
-    // let (caller_private, caller_ether, caller_sol, _caller_nonce) = get_ethereum_caller_credentials(config);
-
-    let (caller_private_eth, caller_ether) = {
-        let private_bytes : [u8; 64] = config.keypair.as_ref().unwrap().to_bytes();
-        let mut sign_arr: [u8;32] = Default::default();
-        sign_arr.clone_from_slice(&private_bytes[..32]);
-        let caller_private = SecretKey::parse(&sign_arr).unwrap();
-        let caller_public = PublicKey::from_secret_key(&caller_private);
-        let caller_ether: H160 = keccak256_h256(&caller_public.serialize()[1..]).into();
-        (caller_private, caller_ether)
-    };
-
-    let (caller_sol, _) = make_solana_program_address(&caller_ether, &config.evm_loader);
-
-    if config.rpc_client.get_account_with_commitment(&caller_sol, CommitmentConfig::confirmed())?.value.is_none() {
-        debug!("Caller account not found");
-        command_create_ether_account(config, &caller_ether, 10_u64.pow(9), 0, token_mint)?;
-    } else {
-        debug!(" Caller account found");
-    }
-
-    // Get caller nonce
-    let (trx_count, caller_ether, caller_token) = get_ether_account_nonce(config, &caller_sol, token_mint)?;
-
-    let (program_id, program_ether, program_nonce, program_token, program_code, program_seed) =
-        get_ethereum_contract_account_credentials(config, &caller_ether, trx_count, token_mint);
-
-    // Check program account to see if partial initialization has occurred
-    let mut instrstruction = create_ethereum_contract_accounts_in_solana(
-        config,
-        &program_id,
-        &program_ether,
-        program_nonce,
-        &program_token,
-        &program_code,
-        &program_seed,
-        program_data.len(),
-        token_mint
-    )?;
-
-    // Create transaction prepared for execution from account
-    let msg = make_deploy_ethereum_transaction(trx_count, &program_data, &caller_private_eth, chain_id);
-
-    // Create holder account (if not exists)
-    let (holder_id, holder_seed) = generate_random_holder_seed();
-    let holder = create_account_with_seed(config, &creator.pubkey(), &creator.pubkey(), &holder_seed, 128*1024_u64)?;
-
-    fill_holder_account(config, &holder, holder_id, &msg)?;
-
-    // Create storage account if not exists
-    let storage = create_storage_account(config)?;
-
-    let (collateral_pool_acc, collateral_pool_index) = get_collateral_pool_account_and_index(config, collateral_pool_base);
-
-    let accounts = vec![
-                        AccountMeta::new(storage, false),
-
-                        AccountMeta::new(creator.pubkey(), true),
-                        AccountMeta::new(collateral_pool_acc, false),
-                        AccountMeta::new(operator_token, false),
-                        AccountMeta::new(caller_token, false),
-                        AccountMeta::new(system_program::id(), false),
-
-                        AccountMeta::new(program_id, false),
-                        AccountMeta::new(program_token, false),
-                        AccountMeta::new(program_code, false),
-                        AccountMeta::new(caller_sol, false),
-                        AccountMeta::new(caller_token, false),
-
-                        AccountMeta::new_readonly(config.evm_loader, false),
-                        AccountMeta::new_readonly(*token_mint, false),
-                        AccountMeta::new_readonly(spl_token::id(), false),
-                        ];
-
-    let mut holder_with_accounts = vec![AccountMeta::new(holder, false)];
-    holder_with_accounts.extend(accounts.clone());
-    // Send trx_from_account_data_instruction
-    {
-        debug!("trx_from_account_data_instruction holder_plus_accounts: {:?}", holder_with_accounts);
-        let trx_from_account_data_instruction = Instruction::new_with_bincode(config.evm_loader,
-                                                                              &(0x16_u8, collateral_pool_index, 0_u64),
-                                                                              holder_with_accounts);
-        instrstruction.push(trx_from_account_data_instruction);
-        send_transaction(config, &instrstruction)?;
-    }
-
-    // Continue while no result
-    loop {
-        let continue_accounts = accounts.clone();
-        debug!("continue continue_accounts: {:?}", continue_accounts);
-        let continue_instruction = Instruction::new_with_bincode(config.evm_loader,
-                                                                 &(0x14_u8, collateral_pool_index, 400_u64),
-                                                                 continue_accounts);
-        let signature = send_transaction(config, &[continue_instruction])?;
-
-        // Check if Continue returned some result
-        let result = config.rpc_client.get_transaction_with_config(
-            &signature,
-            RpcTransactionConfig {
-                commitment: Some(CommitmentConfig::confirmed()),
-                encoding: Some(UiTransactionEncoding::Json),
-            },
-        )?;
-
-        let return_value = parse_transaction_reciept(config, result);
-
-        if let Some(value) = return_value {
-            let (exit_code, data) = value.split_at(1);
-            debug!("exit code {}", exit_code[0]);
-            debug!("return data {}", &hex::encode(data));
-            break;
-        }
-    }
-
-    println!("{}", json!({
-        "programId": format!("{}", program_id),
-        "programToken": format!("{}", program_token),
-        "codeId": format!("{}", program_code),
-        "ethereum": format!("{:?}", program_ether),
-    }));
-    Ok(())
-}
-
-fn command_get_ether_account_data (
-    config: &Config,
-    ether_address: &H160,
-) {
-    match EmulatorAccountStorage::get_account_from_solana(config, ether_address) {
-        Some((acc, balance, code_account)) => {
-            let (solana_address, _solana_nonce) = make_solana_program_address(ether_address, &config.evm_loader);
-            let account_data = AccountData::unpack(&acc.data).unwrap();
-            let account_data = AccountData::get_account(&account_data).unwrap();
-
-            println!("Ethereum address: 0x{}", &hex::encode(&ether_address.as_fixed_bytes()));
-            println!("Solana address: {}", solana_address);
-
-            println!("Account fields");
-            println!("    ether: {}", &account_data.ether);
-            println!("    nonce: {}", &account_data.nonce);
-            println!("    trx_count: {}", &account_data.trx_count);
-            println!("    code_account: {}", &account_data.code_account);
-            println!("    ro_blocked_cnt: {}", &account_data.ro_blocked_cnt);
-            println!("    rw_blocked_acc: {}",
-                     if account_data.rw_blocked_acc.is_some() {
-                         account_data.rw_blocked_acc.unwrap().to_string()
-                     }
-                     else {
-                         "".to_string()
-                     }
-            );
-            println!("    token_account: {}", &account_data.eth_token_account);
-            println!("    token_amount: {}", &balance);
-
-            if let Some(code_account) = code_account {
-                let code_data = AccountData::unpack(&code_account.data).unwrap();
-                let header = AccountData::size(&code_data);
-                let code_data = AccountData::get_contract(&code_data).unwrap();
-
-                println!("Contract fields");
-                println!("    owner: {}", &code_data.owner);
-                println!("    code_size: {}", &code_data.code_size);
-                println!("    code as hex:");
-
-                let code_size = code_data.code_size;
-                let mut offset = header;
-                while offset < ( code_size as usize + header) {
-                    let data_slice = &code_account.data.as_slice();
-                    let remains = if code_size as usize + header - offset > 80 {
-                        80
-                    } else {
-                        code_size as usize + header - offset
-                    };
-
-                    println!("        {}", &hex::encode(&data_slice[offset+header..offset+header+remains]));
-                    offset += remains;
-                }
-            }
-
-
-        },
-        None => {
-            println!("Account not found {}", &ether_address.to_string());
-        }
-    }
-}
-
-fn command_get_storage_at(
-    config: &Config,
-    ether_address: &H160,
-    index: &U256
-) -> CommandResult {
-    match EmulatorAccountStorage::get_account_from_solana(config, ether_address) {
-        Some((acc, _balance, code_account)) => {
-            let account_data = AccountData::unpack(&acc.data)?;
-            let mut code_data = match code_account.as_ref() {
-                Some(code) => code.data.clone(),
-                None => return Err(format!("Account {:#x} is not code account", ether_address).into()),
-            };
-            let contract_data = AccountData::unpack(&code_data)?;
-            let (solana_address, _solana_nonce) = make_solana_program_address(ether_address, &config.evm_loader);
-            let code_data: std::rc::Rc<std::cell::RefCell<&mut [u8]>> = Rc::new(RefCell::new(&mut code_data));
-            let solidity_account = SolidityAccount::new(&solana_address, account_data,
-                                                        Some((contract_data, code_data)));
-            let value = solidity_account.get_storage(index);
-            print!("{:#x}", value);
-            Ok(())
-        },
-        None => {
-            Err(format!("Account not found {:#x}", ether_address).into())
-        }
-    }
-}
-
-fn command_cancel_trx(
-    config: &Config,
-    storage_account: &Pubkey,
-    token_mint: &Pubkey
-) -> CommandResult {
-    let storage = config.rpc_client.get_account_with_commitment(storage_account, CommitmentConfig::processed()).unwrap().value;
-
-    if let Some(acc) = storage {
-        if acc.owner != config.evm_loader {
-            return Err(format!("Invalid owner {} for storage account", acc.owner).into());
-        }
-        let data = AccountData::unpack(&acc.data)?;
-        let data_end = data.size();
-        let storage = if let AccountData::Storage(storage) = data {storage}
-                else {return Err("Not storage account".to_string().into());};
-
-        let keys: Vec<Pubkey> = {
-            info!("{:?}", storage);
-            let accounts_begin = data_end;
-            let accounts_end = accounts_begin + storage.accounts_len * 32;
-            if acc.data.len() < accounts_end {
-                return Err(format!("Accounts data too small: account_data.len()={:?} < end={:?}", acc.data.len(), accounts_end).into());
-            };
-
-            acc.data[accounts_begin..accounts_end].chunks_exact(32).map(Pubkey::new).collect()
-        };
-
-        let (caller_solana, _) = make_solana_program_address(&storage.caller, &config.evm_loader);
-        let (trx_count, _caller_ether, caller_token) = get_ether_account_nonce(config, &caller_solana, token_mint)?;
-
-        let operator = &config.signer.pubkey();
-        let operator_token = spl_associated_token_account::get_associated_token_address(operator, token_mint);
-
-        let mut accounts_meta : Vec<AccountMeta> = vec![
-            AccountMeta::new(*storage_account, false),              // Storage account
-            AccountMeta::new(*operator, true),                      // Operator
-            AccountMeta::new(operator_token, false),                // Operator token
-            AccountMeta::new(caller_token, false),                  // Caller token
-            AccountMeta::new(incinerator::id(), false),             // Incinerator
-            AccountMeta::new_readonly(system_program::id(), false), // System
-        ];
-
-        let system_accounts : Vec<Pubkey> = vec![
-            config.evm_loader,
-            *token_mint,
-            spl_token::id(),
-            spl_associated_token_account::id(),
-            sysvar::rent::id(),
-            incinerator::id(),
-            system_program::id(),
-            sysvar::instructions::id(),
-        ];
-
-        for key in keys {
-            let writable = if system_accounts.contains(&key) {false} else {
-                let acc = config.rpc_client.get_account_with_commitment(&key, CommitmentConfig::processed()).unwrap().value;
-                if let Some(acc) = acc {
-                    if acc.owner == config.evm_loader {
-                        matches!(AccountData::unpack(&acc.data)?, AccountData::Account(_))
-                    } else {
-                        false
-                    }
-                } else {false}
-            };
-
-            if writable {
-                accounts_meta.push(AccountMeta::new(key, false));
-            } else {
-                accounts_meta.push(AccountMeta::new_readonly(key, false));
-            }
-        }
-        for meta in &accounts_meta {
-            info!("\t{:?}", meta);
-        }
-
-        let instruction = Instruction::new_with_bincode(config.evm_loader, &(21_u8, trx_count), accounts_meta);
-        send_transaction(config, &[instruction])?;
-
-    } else {
-        return Err(format!("Account not found {}", &storage_account.to_string()).into());
-    }
-    Ok(())
-}
-
-fn read_elf_parameters(
-        _config: &Config,
-        program_data: &[u8],
-    )-> HashMap<String, String>{
-    let mut result = HashMap::new();
-    let elf = goblin::elf::Elf::parse(program_data).expect("Unable to parse ELF file");
-    elf.dynsyms.iter().for_each(|sym| {
-        let name = String::from(&elf.dynstrtab[sym.st_name]);
-        if name.starts_with("NEON")
-        {
-            let end = program_data.len();
-            let from = usize::try_from(sym.st_value).unwrap_or_else(|_| panic!("Unable to cast usize from u64:{:?}", sym.st_value));
-            let to = usize::try_from(sym.st_value + sym.st_size).unwrap_or_else(|err| panic!("Unable to cast usize from u64:{:?}. Error: {}", sym.st_value + sym.st_size, err));
-            if to < end && from < end {
-                let buf = &program_data[from..to];
-                let value = std::str::from_utf8(buf).unwrap();
-                result.insert(name, String::from(value));
-            }
-            else {
-                panic!("{} is out of bounds", name);
-            }
-        }
-    });
-
-    result
-}
-
-
-fn read_elf_parameters_from_account(config: &Config) -> Result<HashMap<String, String>, Error> {
-    let account = config.rpc_client
-        .get_account_with_commitment(&config.evm_loader, config.commitment)?
-        .value.ok_or(format!("Unable to find the account {}", &config.evm_loader))?;
-
-    if account.owner == bpf_loader::id() || account.owner == bpf_loader_deprecated::id() {
-        Ok(read_elf_parameters(config, &account.data))
-    } else if account.owner == bpf_loader_upgradeable::id() {
-        if let Ok(UpgradeableLoaderState::Program {
-                      programdata_address,
-                  }) = account.state()
-        {
-            let programdata_account = config.rpc_client
-                .get_account_with_commitment(&programdata_address, config.commitment)?
-                .value.ok_or(format!(
-                "Failed to find associated ProgramData account {} for the program {}",
-                programdata_address, &config.evm_loader))?;
-
-            if let Ok(UpgradeableLoaderState::ProgramData { .. }) = programdata_account.state() {
-                let offset =
-                    UpgradeableLoaderState::programdata_data_offset().unwrap_or(0);
-                let program_data = &programdata_account.data[offset..];
-                Ok(read_elf_parameters(config, program_data))
-            } else {
-                Err(
-                    format!("Invalid associated ProgramData account {} found for the program {}",
-                            programdata_address, &config.evm_loader)
-                        .into(),
-                )
-            }
-
-        } else if let Ok(UpgradeableLoaderState::Buffer { .. }) = account.state() {
-            let offset = UpgradeableLoaderState::buffer_data_offset().unwrap_or(0);
-            let program_data = &account.data[offset..];
-            Ok(read_elf_parameters(config, program_data))
-        } else {
-            Err(format!(
-                "{} is not an upgradeble loader buffer or program account",
-                &config.evm_loader
-            )
-                .into())
-        }
-    } else {
-        Err(format!("{} is not a BPF program", &config.evm_loader).into())
-    }
-
-}
-
-
-fn print_elf_parameters(params: &HashMap<String, String>){
-    for (key, value) in params {
-        println!("{}={}", key, value);
-    }
-}
-
-
-fn read_program_data_from_file(config: &Config,
-                               program_location: &str) -> CommandResult {
-    let program_data = read_program_data(program_location)?;
-    let program_data = &program_data[..];
-    let elf_params = read_elf_parameters(config, program_data);
-    print_elf_parameters(&elf_params);
-    Ok(())
-}
-
-
-fn read_program_data_from_account(config: &Config) {
-    let elf_params = read_elf_parameters_from_account(config).unwrap();
-    print_elf_parameters(&elf_params);
-}
-
-
-fn command_neon_elf(
-    config: &Config,
-    program_location: Option<&str>,
-) -> CommandResult {
-    program_location.map_or_else(
-        || {read_program_data_from_account(config); Ok(())},
-        |program_location| read_program_data_from_file(config, program_location),
-    )
-}
-
-fn command_update_valids_table(
-    config: &Config,
-    ether_address: &H160,
-) -> CommandResult {
-    let account_data = if let Some((account, _, _)) = EmulatorAccountStorage::get_account_from_solana(config, ether_address) {
-        AccountData::unpack(&account.data)?
-    } else {
-        return Err(format!("Account not found {:#x}", ether_address).into());
-    };
-
-    let code_account = account_data.get_account()?.code_account;
-    if code_account == Pubkey::new_from_array([0_u8; 32]) {
-        return Err(format!("Code account not found {:#x}", ether_address).into());
-    }
-
-    let instruction = Instruction::new_with_bincode(
-        config.evm_loader,
-        &(23),
-        vec![AccountMeta::new(code_account, false)]
-    );
-
-    let finalize_message = Message::new(&[instruction], Some(&config.signer.pubkey()));
-    let (blockhash, fee_calculator) = config.rpc_client.get_recent_blockhash()?;
-
-    check_account_for_fee(
-        &config.rpc_client,
-        &config.signer.pubkey(),
-        &fee_calculator,
-        &finalize_message)?;
-
-    let mut finalize_tx = Transaction::new_unsigned(finalize_message);
-    finalize_tx.try_sign(&[&*config.signer], blockhash)?;
-    debug!("signed: {:x?}", finalize_tx);
-
-    config.rpc_client.send_and_confirm_transaction_with_spinner(&finalize_tx)?;
-
-    Ok(())
 }
 
 fn make_clean_hex(in_str: &str) -> &str {
@@ -1743,7 +717,7 @@ fn main() {
 
         let context: LogContext =
             app_matches.value_of("logging_ctx")
-                .map(|ctx| serde_json::from_str(ctx).unwrap() )
+                .map(|ctx| LogContext::new(ctx.to_string()) )
                 .unwrap_or_default();
         logs::init(context).unwrap();
 
@@ -1829,15 +803,15 @@ fn main() {
 
                 let token_mint = pubkey_of(arg_matches, "token_mint")
                     .unwrap_or_else(|| {
-                        let elf_params = read_elf_parameters_from_account(&config).unwrap();
+                        let elf_params = get_neon_elf::read_elf_parameters_from_account(&config).unwrap();
                         Pubkey::from_str(elf_params.get("NEON_TOKEN_MINT").unwrap()).unwrap()
                     });
-                command_emulate(&config, contract, sender, data, value, &token_mint)
+                emulate::command_emulate(&config, contract, sender, data, value, &token_mint)
             }
             ("create-program-address", Some(arg_matches)) => {
                 let ether = h160_of(arg_matches, "seed").unwrap();
 
-                command_create_program_address(&config, &ether);
+                create_program_address::command_create_program_address(&config, &ether);
 
                 Ok(())
             }
@@ -1848,12 +822,12 @@ fn main() {
 
                 let token_mint = pubkey_of(arg_matches, "token_mint")
                     .unwrap_or_else(|| {
-                        let elf_params = read_elf_parameters_from_account(&config).unwrap();
+                        let elf_params = get_neon_elf::read_elf_parameters_from_account(&config).unwrap();
                         Pubkey::from_str(elf_params.get("NEON_TOKEN_MINT").unwrap()).unwrap()
                     });
 
 
-                command_create_ether_account(&config, &ether, lamports, space, &token_mint)
+                create_ether_account::command_create_ether_account(&config, &ether, lamports, space, &token_mint)
             }
             ("deploy", Some(arg_matches)) => {
                 let program_location = arg_matches.value_of("program_location").unwrap().to_string();
@@ -1862,14 +836,14 @@ fn main() {
 
                 let token_mint = pubkey_of(arg_matches, "token_mint")
                     .unwrap_or_else(|| {
-                        elf_params = read_elf_parameters_from_account(&config).unwrap();
+                        elf_params = get_neon_elf::read_elf_parameters_from_account(&config).unwrap();
                         Pubkey::from_str(elf_params.get("NEON_TOKEN_MINT").unwrap()).unwrap()
                     });
 
                 let collateral_pool_base = pubkey_of(arg_matches, "collateral_pool_base")
                     .unwrap_or_else(|| {
                         if elf_params.is_empty(){
-                            elf_params = read_elf_parameters_from_account(&config).unwrap();
+                            elf_params = get_neon_elf::read_elf_parameters_from_account(&config).unwrap();
                         }
                         Pubkey::from_str(elf_params.get("NEON_POOL_BASE").unwrap()).unwrap()
                     });
@@ -1877,16 +851,16 @@ fn main() {
                 let chain_id = value_of(arg_matches, "chain_id")
                     .unwrap_or_else(|| {
                         if elf_params.is_empty(){
-                            elf_params = read_elf_parameters_from_account(&config).unwrap();
+                            elf_params = get_neon_elf::read_elf_parameters_from_account(&config).unwrap();
                         }
                         u64::from_str(elf_params.get("NEON_CHAIN_ID").unwrap()).unwrap()
                     });
-                command_deploy(&config, &program_location, &token_mint, &collateral_pool_base, chain_id)
+                deploy::command_deploy(&config, &program_location, &token_mint, &collateral_pool_base, chain_id)
             }
             ("get-ether-account-data", Some(arg_matches)) => {
                 let ether = h160_of(arg_matches, "ether").unwrap();
 
-                command_get_ether_account_data(&config, &ether);
+                get_ether_account_data::command_get_ether_account_data(&config, &ether);
 
                 Ok(())
             }
@@ -1895,26 +869,26 @@ fn main() {
 
                 let token_mint = pubkey_of(arg_matches, "token_mint")
                     .unwrap_or_else(|| {
-                        let elf_params = read_elf_parameters_from_account(&config).unwrap();
+                        let elf_params = get_neon_elf::read_elf_parameters_from_account(&config).unwrap();
                         Pubkey::from_str(elf_params.get("NEON_TOKEN_MINT").unwrap()).unwrap()
                     });
-                command_cancel_trx(&config, &storage_account, &token_mint)
+                cancel_trx::command_cancel_trx(&config, &storage_account, &token_mint)
             }
             ("neon-elf-params", Some(arg_matches)) => {
                 let program_location = arg_matches.value_of("program_location");
 
-                command_neon_elf(&config, program_location)
+                get_neon_elf::command_neon_elf(&config, program_location)
             }
             ("get-storage-at", Some(arg_matches)) => {
                 let contract_id = h160_of(arg_matches, "contract_id").unwrap();
                 let index = u256_of(arg_matches, "index").unwrap();
 
-                command_get_storage_at(&config, &contract_id, &index)
+                get_storage_at::command_get_storage_at(&config, &contract_id, &index)
             }
             ("update-valids-table", Some(arg_matches)) => {
                 let contract_id = h160_of(arg_matches, "contract_id").unwrap();
 
-                command_update_valids_table(&config, &contract_id)
+                update_valids_table::command_update_valids_table(&config, &contract_id)
             }
             _ => unreachable!(),
         };
